Added Photo model, Photos controller, updated nav integration tests
The Photo model and its controller were created and connected to the site structure through a redirect to first Photo of a Project from that Project's index page. Only the navigation of Project#show has been tested (by simply reusing the same test logic as the pre-redirect tests of the previous commit).